+++ src/module/general/ReadEnsight/EnGoldGeoBIN.cpp
@@ -843,9 +843,9 @@ EnGoldGeoBIN::parseForParts()
 int
 EnGoldGeoBIN::allocateMemory()
 {
-    int totNumCoords(0);
-    int totNumEle(0);
-    int totNumCorners(0);
+    uint64_t totNumCoords(0);
+    uint64_t totNumEle(0);
+    uint64_t totNumCorners(0);
 
     if (!ens->masterPL_.empty())
     {
